Recipes-json file and items-json file are created
This script creates a json file with the attributes needed for a recipe. The id, the name, the reagents and what it creates, the official blizzard recipe endpoint doesn't give you the reagents or what it creates, the only thing that ocurred to me, was to use information from wowhead which had to be scrapped and then batch processed to create this.
The second file created is a list with every item required for all these recipes, an unsorted - non duplicated list.